seg fault sto 16

+++ JoinArray.h
@@ -21,8 +21,8 @@ public:
     JoinArray(Relations *r);
     void set_currentColumn(int column); //antikatastithike apo tin setrel(se diaforetika simia)
     //giati o pinakas 1 edo mporei na einai sti thesi 0
-    uint64_t get_value(int i);
-    void set_value(int i,uint64_t val);
+    uint64_t get_value(uint64_t i);
+    void set_value(uint64_t i,uint64_t val);
     void insert_row(uint64_t i,uint64_t *row);
     void update_array(list *results,int id);
     void update_array(list *results,JoinArray *array2);
